Added a call to EndAttack on Kirby's current attack before he creates a new one in EndSwallow, it was causing sound leaks. That system probably needs some housekeeping later anyway.
As a failsafe, the mute key (M) now stops all sounds when you press it. It's probably more likely that we'd run into a sound leak and need to quickly stop it during our demo than it is we'd ever want mute to not interrupt existing sounds, so I figure this is probably for the best to have.